rename headers to match convention, move plugin spec back to core for later use
