+++ src/Core/CMakeLists.txt
@@ -5,8 +5,9 @@ SPDX-License-Identifier: GPL-2.0-or-later
 ]===]
 
 add_library(Mupen64RR.Core.Headers INTERFACE
-    "include/m64rr/core_types.hpp"
-    "include/m64rr/core_api.hpp"
+    "include/m64rr/API.hpp"
+    "include/m64rr/Plugin.hpp"
+    "include/m64rr/Types.hpp"
 )
 set_target_properties(Mupen64RR.Core.Headers PROPERTIES
     CXX_STANDARD 23
