+++ src/Plugins.Win32.TASAudio/Main.hpp
@@ -7,7 +7,7 @@
 #pragma once
 #include "Config.hpp"
 #include "SDLBackend.hpp"
-#include <Views.Win32/M64RRSpec.h>
+#include <m64rr/Plugin.hpp>
 #include <VersionNameHelpers.hpp>
 
 #define PLUGIN_NAME VERSION_NAME_HELPER_GEN_NAME(L"TAS Audio")
