+++ src/Plugins.Win32.TASVideo/glN64.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Views.Win32/M64RRSpec.h>
+#include <m64rr/Plugin.hpp>
 
 struct TASVideoContext
 {
