+++ src/Views.Win32/CMakeLists.txt
@@ -6,7 +6,6 @@ SPDX-License-Identifier: GPL-2.0-or-later
 
 add_library(Mupen64RR.Views.Win32.Headers INTERFACE
     "include/Views.Win32/ZESpec.h"
-    "include/Views.Win32/M64RRSpec.h"
 )
 set_target_properties(Mupen64RR.Views.Win32.Headers PROPERTIES
     CXX_STANDARD 23
