+++ src/Views.Win32/stdafx.h
@@ -11,7 +11,7 @@
 #define SPDLOG_LEVEL_NAMES {"🔍", "🪲", "ℹ️", "⚠️", "❌", "💥", ""}
 
 #include <CommonPCH.hpp>
-#include <m64rr/core_api.hpp>
+#include <m64rr/API.hpp>
 
 #pragma warning(push, 0)
 extern "C"
